+++ src/mailman/rest/wsgiapp.py
@@ -25,10 +25,9 @@ from falcon import API, HTTPUnauthorized
 from falcon.routing import map_http_methods, set_default_responders
 from mailman.config import config
 from mailman.database.transaction import transactional
+from mailman.rest.helpers import bad_request
 from mailman.rest.root import Root
 from public import public
-from wsgiref.simple_server import (
-    WSGIRequestHandler, WSGIServer, make_server as wsgi_server)
 
 
 log = logging.getLogger('mailman.http')
@@ -39,44 +38,6 @@ EMPTYSTRING = ''
 REALM = 'mailman3-rest'
 UTF8 = 'utf-8'
 WILDCARD_ACCEPT_HEADER = '*/*'
-
-
-class AdminWSGIServer(WSGIServer):
-    """Server class that integrates error handling with our log files."""
-
-    def handle_error(self, request, client_address):
-        # Interpose base class method so that the exception gets printed to
-        # our log file rather than stderr.
-        log.exception('REST server exception during request from %s',
-                      client_address)
-
-
-class StderrLogger:
-    def __init__(self):
-        self._buffer = []
-
-    def write(self, message):
-        self._buffer.append(message)
-
-    def flush(self):
-        self._buffer.insert(0, 'REST request handler error:\n')
-        log.error(EMPTYSTRING.join(self._buffer))
-        self._buffer = []
-
-
-class AdminWebServiceWSGIRequestHandler(WSGIRequestHandler):
-    """Handler class which just logs output to the right place."""
-
-    default_request_version = 'HTTP/1.1'
-
-    def log_message(self, format, *args):
-        """See `BaseHTTPRequestHandler`."""
-        log.info('%s - - %s', self.address_string(), format % args)
-
-    def get_stderr(self):
-        # Return a fake stderr object that will actually write its output to
-        # the log file.
-        return StderrLogger()
 
 
 class Middleware:
@@ -103,6 +64,20 @@ class Middleware:
                 '401 Unauthorized',
                 'REST API authorization failed',
                 challenges=[realm])
+
+
+def handle_ValueError(exc, request, response, params):
+    """Handle ValueErrors in API code to return HTTPBadRequest.
+
+    ValueErrors are raised often by Validator and should not return a 500 error
+    resposne to the client.  This is a stop-gap for 500 errors due to
+    ValueErrors, it is recommended that they be handled at the call-site,
+    instead of here.
+    """
+    # Only handle ValueError, raise anything else right back.
+    if not isinstance(exc, ValueError):
+        raise exc
+    bad_request(response, str(exc))
 
 
 class ObjectRouter:
@@ -221,19 +196,6 @@ class RootedAPI(API):
 @public
 def make_application():
     """Return a callable WSGI application object."""
-    return RootedAPI(Root())
-
-
-@public
-def make_server():
-    """Create the Mailman REST server.
-
-    Use this if you just want to run Mailman's wsgiref-based REST server.
-    """
-    host = config.webservice.hostname
-    port = int(config.webservice.port)
-    server = wsgi_server(
-        host, port, make_application(),
-        server_class=AdminWSGIServer,
-        handler_class=AdminWebServiceWSGIRequestHandler)
-    return server
+    app = RootedAPI(Root())
+    app.add_error_handler(ValueError, handle_ValueError)
+    return app